+++ example/integration_test/src/tests/unit/src/view/test_app_state.dart
@@ -19,11 +19,11 @@ void testAppState(WidgetTester tester) {
 
   bool? use = appState.useCupertino;
 
-  use = appState.isCupertino;
+  use = appState.useCupertino;
 
   use = appState.useMaterial;
 
-  use = appState.isMaterial;
+  use = appState.useMaterial;
 
   appState.onHome();